Collapse hspd12_policy into aal3_policy
