+++ app/controllers/concerns/mfa_setup_concern.rb
@@ -5,7 +5,7 @@ module MfaSetupConcern
     authenticate_user!(force: true)
     return if user_fully_authenticated?
     return unless MfaPolicy.new(current_user, session).two_factor_enabled?
-    return if hspd12_policy.piv_cac_setup_required?
+    return if aal3_policy.piv_cac_only_setup_required?
     redirect_to user_two_factor_authentication_url
   end
 end
