+++ app/controllers/two_factor_authentication/piv_cac_verification_controller.rb
@@ -69,7 +69,7 @@ module TwoFactorAuthentication
     def piv_cac_view_data
       {
         two_factor_authentication_method: two_factor_authentication_method,
-        hide_fallback_question: hspd12_policy.piv_cac_required?,
+        hide_fallback_question: aal3_policy.piv_cac_only_required?,
         user_email: current_user.email_addresses.take.email,
       }.merge(generic_data)
     end
