+++ app/decorators/mfa_context.rb
@@ -60,7 +60,7 @@ class MfaContext
   end
 
   def two_factor_configurations
-    return piv_cac_configurations if piv_cac_required?
+    return piv_cac_configurations if piv_cac_only_required?
     phone_configurations + webauthn_configurations + backup_code_configurations +
       piv_cac_configurations + auth_app_configurations
   end
@@ -96,8 +96,8 @@ class MfaContext
 
   private
 
-  def piv_cac_required?
-    Hspd12Policy.new(session: @session, user: user).piv_cac_required?
+  def piv_cac_only_required?
+    AAL3Policy.new(session: @session, user: user).piv_cac_only_required?
   end
 
   def personal_key_method_count
