+++ app/presenters/two_factor_options_presenter.rb
@@ -12,12 +12,12 @@ class TwoFactorOptionsPresenter
   end
 
   def icon
-    return 'icon-lock-alert-important.svg' if piv_cac_required? ||
+    return 'icon-lock-alert-important.svg' if piv_cac_only_required? ||
                                               (aal3_only? && mfa_policy.two_factor_enabled?)
   end
 
   def heading
-    if piv_cac_required?
+    if piv_cac_only_required?
       t('two_factor_authentication.two_factor_hspd12_choice')
     elsif aal3_only? && mfa_policy.two_factor_enabled?
       t('two_factor_authentication.two_factor_aal3_choice')
@@ -27,7 +27,7 @@ class TwoFactorOptionsPresenter
   end
 
   def intro
-    if piv_cac_required?
+    if piv_cac_only_required?
       t('two_factor_authentication.two_factor_hspd12_choice_intro')
     elsif aal3_only?
       t('two_factor_authentication.two_factor_aal3_choice_intro')
@@ -37,7 +37,7 @@ class TwoFactorOptionsPresenter
   end
 
   def show_security_level?
-    !(piv_cac_required? || (aal3_only? && mfa_policy.two_factor_enabled?))
+    !(piv_cac_only_required? || (aal3_only? && mfa_policy.two_factor_enabled?))
   end
 
   private
@@ -48,22 +48,22 @@ class TwoFactorOptionsPresenter
   end
 
   def webauthn_option
-    return [] if piv_cac_required?
+    return [] if piv_cac_only_required?
     [TwoFactorAuthentication::WebauthnSelectionPresenter.new]
   end
 
   def phone_options
-    return [] if piv_cac_required? || aal3_only? || FeatureManagement.hide_phone_mfa_signup?
+    return [] if piv_cac_only_required? || aal3_only? || FeatureManagement.hide_phone_mfa_signup?
     [TwoFactorAuthentication::PhoneSelectionPresenter.new]
   end
 
   def totp_option
-    return [] if piv_cac_required? || aal3_only?
+    return [] if piv_cac_only_required? || aal3_only?
     [TwoFactorAuthentication::AuthAppSelectionPresenter.new]
   end
 
   def backup_code_option
-    return [] if piv_cac_required? || aal3_only?
+    return [] if piv_cac_only_required? || aal3_only?
     [TwoFactorAuthentication::BackupCodeSelectionPresenter.new]
   end
 
@@ -71,8 +71,8 @@ class TwoFactorOptionsPresenter
     DeviceDetector.new(@user_agent)&.device_type == 'desktop'
   end
 
-  def piv_cac_required?
-    hspd12_policy.piv_cac_required?
+  def piv_cac_only_required?
+    aal3_policy.piv_cac_only_required?
   end
 
   def aal3_only?
@@ -81,9 +81,5 @@ class TwoFactorOptionsPresenter
 
   def mfa_policy
     @mfa_policy ||= MfaPolicy.new(@user)
-  end
-
-  def hspd12_policy
-    @hspd12_policy ||= Hspd12Policy.new(session: @session, user: @user)
   end
 end
